Lecture-10-LangGraph-Checkpointing/assignment/pyproject.toml
[project]
name = "assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langchain-qdrant>=0.2.0",
    "openai>=1.86.0",
    "pyaudio>=0.2.14",
    "speechrecognition>=3.14.3",
]
